Moved select component to components folder and fixed react import errors
