+++ src/js/shot-generator-new/components/SimpleGrid/index.js
@@ -1,3 +1,4 @@
+import React from 'react'
 const elementStyle = {
     position:"absolute", 
     height:{},
